extracted logic in multiple files
