--- src/bp/nlu-core/entities/duckling_extractor.ts
@@ -1,416 +1,0 @@
-import Axios, { AxiosInstance } from 'axios'
-import retry from 'bluebird-retry'
-import { NLU } from 'botpress/sdk'
-import { ensureFile, pathExists, readJSON, writeJson } from 'fs-extra'
-import httpsProxyAgent from 'https-proxy-agent'
-import _ from 'lodash'
-import lru from 'lru-cache'
-import ms from 'ms'
-import sizeof from 'object-sizeof'
-import path from 'path'
-
-import { extractPattern } from '../tools/patterns-utils'
-import { SPACE } from '../tools/token-utils'
-import { EntityExtractionResult, SystemEntityExtractor } from '../typings'
-
-interface DucklingParams {
-  tz: string
-  refTime: number
-  lang: string
-}
-
-interface KeyedItem {
-  input: string
-  idx: number
-  entities?: EntityExtractionResult[]
-}
-
-/**
- * duckling return for "prompt-date next-week"
- * {
- *   "body": "next week",
- *   "start": 12,
- *   "value": {
- *     "values": [
- *       {
- *         "value": "2020-10-12T00:00:00.000-04:00",
- *         "grain": "week",
- *         "type": "value"
- *       }
- *     ],
- *     "value": "2020-10-12T00:00:00.000-04:00",
- *     "grain": "week",
- *     "type": "value"
- *   },
- *   "end": 21,
- *   "dim": "time",
- *   "latent": false
- * }
- *
- *
- * OR
- *
- *
- * {
- *  "values": [
- *    {
- *      "to": {
- *        "value": "2020-10-08T00:00:00.000-04:00",
- *        "grain": "hour"
- *      },
- *      "from": {
- *        "value": "2020-10-07T18:00:00.000-04:00",
- *        "grain": "hour"
- *      },
- *      "type": "interval"
- *    }
- *  ],
- *  "to": {
- *    "value": "2020-10-08T00:00:00.000-04:00",
- *    "grain": "hour"
- *  },
- *  "from": {
- *    "value": "2020-10-07T18:00:00.000-04:00",
- *    "grain": "hour"
- *  },
- *  "type": "interval"
- * }
- */
-interface DucklingReturn {
-  start: number
-  end: number
-  dim: string
-  body: string
-  value: DucklingValue
-}
-interface DucklingValue {
-  normalized?: ValueUnit
-  values?: (ValueGrain | IntervalValueGrain)[]
-  value?: string
-  unit?: string
-  grain?: string
-}
-
-interface DucklingTime extends DucklingReturn {
-  dim: 'time'
-  value: DucklingTimeValue
-}
-type DucklingTimeValue = ValueGrain | DucklingTimeValues
-type DucklingTimeValues = { values: (ValueGrain | IntervalValueGrain)[] }
-
-interface DucklingDuration extends DucklingReturn {
-  dim: 'duration'
-  value: { normalized: ValueUnit }
-}
-
-interface ValueUnit {
-  value: string
-  unit: string
-  type: 'value'
-}
-
-interface ValueGrain {
-  value: string
-  grain: string
-  type: 'value'
-}
-
-interface IntervalValueGrain {
-  from: ValueGrain
-  to: ValueGrain
-  type: 'interval'
-}
-
-// duckling type guards
-const _isDuration = (duck: DucklingReturn): duck is DucklingDuration => {
-  return duck.dim === 'duration'
-}
-
-const _isTime = (duck: DucklingReturn): duck is DucklingTime => {
-  return duck.dim === 'time'
-}
-
-const _isTimeValues = (duckValue: DucklingTimeValue): duckValue is DucklingTimeValues => {
-  return !!(duckValue as DucklingTimeValues).values
-}
-
-const _isValueGrain = (duckValue: DucklingTimeValue): duckValue is ValueGrain => {
-  return !!((duckValue as ValueGrain).value && (duckValue as ValueGrain).grain)
-}
-
-const _isValueUnit = (duckValue: DucklingValue): duckValue is ValueUnit => {
-  return !!((duckValue as ValueUnit).value && (duckValue as ValueUnit).unit)
-}
-
-const _isIntervalValueGrain = (duckValue: ValueGrain | IntervalValueGrain): duckValue is IntervalValueGrain => {
-  return duckValue.type === 'interval'
-}
-
-export const JOIN_CHAR = `::${SPACE}::`
-const BATCH_SIZE = 10
-const DISABLED_MSG = `, so it will be disabled.
-For more information (or if you want to self-host it), please check the docs at
-https://botpress.com/docs/build/nlu/#system-entities
-`
-const DUCKLING_ENTITIES = [
-  'amountOfMoney',
-  'distance',
-  'duration',
-  'email',
-  'number',
-  'ordinal',
-  'phoneNumber',
-  'quantity',
-  'temperature',
-  'time',
-  'url',
-  'volume'
-]
-
-const RETRY_POLICY = { backoff: 2, max_tries: 3, timeout: 500 }
-const CACHE_PATH = path.join(process.APP_DATA_PATH || '', 'cache', 'sys_entities.json')
-
-// Further improvements:
-// 1 - Duckling entity interface
-// 3- in _extractBatch, shift results ==> don't walk whole array n times (nlog(n) vs n2)
-
-export class DucklingEntityExtractor implements SystemEntityExtractor {
-  public static enabled: boolean
-  public static client: AxiosInstance
-
-  private static _cache: lru<string, EntityExtractionResult[]>
-  private _cacheDumpEnabled = true
-
-  constructor(private readonly logger?: NLU.Logger) {}
-
-  public static get entityTypes(): string[] {
-    return DucklingEntityExtractor.enabled ? DUCKLING_ENTITIES : []
-  }
-
-  public static async configure(enabled: boolean, url: string, logger?: NLU.Logger) {
-    if (enabled) {
-      const proxyConfig = process.PROXY ? { httpsAgent: new httpsProxyAgent(process.PROXY) } : {}
-      this.client = Axios.create({
-        baseURL: url,
-        headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-        ...proxyConfig
-      })
-
-      try {
-        await retry(async () => {
-          const { data } = await this.client.get('/')
-          if (data !== 'quack!') {
-            return logger && logger.warning(`Bad response from Duckling server ${DISABLED_MSG}`)
-          }
-          this.enabled = true
-        }, RETRY_POLICY)
-      } catch (err) {
-        logger && logger.warning(`Couldn't reach the Duckling server ${DISABLED_MSG}`, err)
-      }
-
-      this._cache = new lru<string, EntityExtractionResult[]>({
-        length: (val: any, key: string) => sizeof(val) + sizeof(key),
-        max:
-          1000 * // n bytes per entity
-          2 * // entities per utterance
-          10 * // n utterances per intent
-          100 * // n intents per bot
-          50 // n bots
-        // ~ 100 mb
-      })
-
-      await this._restoreCache()
-    }
-  }
-
-  private static async _restoreCache() {
-    try {
-      if (await pathExists(CACHE_PATH)) {
-        const dump = await readJSON(CACHE_PATH)
-        if (dump) {
-          this._cache.load(dump)
-        }
-      }
-    } catch (err) {
-      console.error('could not load duckling cache')
-    }
-  }
-
-  public async extractMultiple(
-    inputs: string[],
-    lang: string,
-    useCache?: boolean
-  ): Promise<EntityExtractionResult[][]> {
-    if (!DucklingEntityExtractor.enabled) {
-      return Array(inputs.length).fill([])
-    }
-    const options = {
-      lang,
-      tz: this._getTz(),
-      refTime: Date.now()
-    }
-
-    const [cached, toFetch] = inputs.reduce(
-      ([cached, toFetch], input, idx) => {
-        if (useCache && DucklingEntityExtractor._cache.has(input)) {
-          const entities = DucklingEntityExtractor._cache.get(input)
-          return [[...cached, { input, idx, entities }], toFetch]
-        } else {
-          return [cached, [...toFetch, { input, idx }]]
-        }
-      },
-      [[], []] as [KeyedItem[], KeyedItem[]]
-    )
-
-    const chunks = _.chunk(toFetch, BATCH_SIZE)
-    const batchedRes = await Promise.mapSeries(chunks, c => this._extractBatch(c, options))
-
-    return _.chain(batchedRes)
-      .flatten()
-      .concat(cached)
-      .orderBy('idx')
-      .map(x => x.entities!)
-      .value()
-  }
-
-  public async extract(input: string, lang: string, useCache?: boolean): Promise<EntityExtractionResult[]> {
-    return (await this.extractMultiple([input], lang, useCache))[0]
-  }
-
-  private async _dumpCache() {
-    try {
-      await ensureFile(CACHE_PATH)
-      await writeJson(CACHE_PATH, DucklingEntityExtractor._cache.dump())
-    } catch (err) {
-      this.logger?.error(`could not persist system entities cache, error ${err.message}`, err)
-      this._cacheDumpEnabled = false
-    }
-  }
-
-  private _onCacheChanged = _.debounce(async () => {
-    if (this._cacheDumpEnabled) {
-      await this._dumpCache()
-    }
-  }, ms('10s'))
-
-  private async _extractBatch(batch: KeyedItem[], params: DucklingParams): Promise<KeyedItem[]> {
-    if (_.isEmpty(batch)) {
-      return []
-    }
-    // trailing JOIN_CHAR so we have n joints and n examples
-    const strBatch = batch.map(x => x.input)
-    const concatBatch = strBatch.join(JOIN_CHAR) + JOIN_CHAR
-    const batchEntities = await this._fetchDuckling(concatBatch, params)
-    const splitLocations = extractPattern(concatBatch, new RegExp(JOIN_CHAR)).map(v => v.sourceIndex)
-    const entities = splitLocations.map((to, idx, locs) => {
-      const from = idx === 0 ? 0 : locs[idx - 1] + JOIN_CHAR.length
-      return batchEntities
-        .filter(e => e.start >= from && e.end <= to)
-        .map(e => ({
-          ...e,
-          start: e.start - from,
-          end: e.end - from
-        }))
-    })
-    await this._cacheBatchResults(strBatch, entities)
-
-    return batch.map((batchItm, i) => ({ ...batchItm, entities: entities[i] }))
-  }
-
-  private async _fetchDuckling(text: string, { lang, tz, refTime }: DucklingParams): Promise<EntityExtractionResult[]> {
-    try {
-      return await retry(async () => {
-        const { data } = await DucklingEntityExtractor.client.post(
-          '/parse',
-          `lang=${lang}&text=${encodeURI(text)}&reftime=${refTime}&tz=${tz}`
-        )
-
-        if (!_.isArray(data)) {
-          throw new Error('Unexpected response from Duckling. Expected an array.')
-        }
-
-        return data.map(this._mapDuckToEntity.bind(this))
-      }, RETRY_POLICY)
-    } catch (err) {
-      const error = err.response ? err.response.data : err
-      this.logger && this.logger.warning('Error extracting duckling entities', error)
-      return []
-    }
-  }
-
-  private async _cacheBatchResults(inputs: string[], results: EntityExtractionResult[][]) {
-    _.zipWith(inputs, results, (input, entities) => ({ input, entities })).forEach(({ input, entities }) => {
-      DucklingEntityExtractor._cache.set(input, entities)
-    })
-
-    await this._onCacheChanged()
-  }
-
-  private _getTz(): string {
-    return Intl.DateTimeFormat().resolvedOptions().timeZone
-  }
-
-  private _mapDuckToEntity(duckEnt: DucklingReturn): EntityExtractionResult {
-    const dimensionData = this._getUnitAndValue(duckEnt)
-    return {
-      confidence: 1,
-      start: duckEnt.start,
-      end: duckEnt.end,
-      type: duckEnt.dim,
-      value: dimensionData.value,
-      metadata: {
-        extractor: 'system',
-        source: duckEnt.body,
-        entityId: `system.${duckEnt.dim}`,
-        unit: dimensionData.unit
-      }
-    } as EntityExtractionResult
-  }
-
-  private _getUnitAndValue(duck: DucklingReturn): ValueUnit {
-    if (_isDuration(duck)) {
-      return duck.value.normalized
-    }
-
-    if (_isTime(duck)) {
-      if (_isValueGrain(duck.value)) {
-        return {
-          value: duck.value.value,
-          unit: duck.value.grain,
-          type: 'value'
-        }
-      }
-
-      if (_isTimeValues(duck.value) && duck.value.values.length) {
-        const first = duck.value.values[0]
-        if (_isIntervalValueGrain(first)) {
-          return {
-            value: first.from.value,
-            unit: first.from.grain,
-            type: 'value'
-          }
-        }
-
-        return {
-          value: first.value,
-          unit: first.grain,
-          type: 'value'
-        }
-      }
-    }
-
-    if (_isValueUnit(duck.value)) {
-      return {
-        value: duck.value.value,
-        unit: duck.value.unit,
-        type: 'value'
-      }
-    }
-
-    return {
-      value: '',
-      unit: '',
-      type: 'value'
-    }
-  }
-}
